[US5045] make admin api endpoints tolerant of multiple realms per edorg

+++ sli/api/src/main/java/org/slc/sli/api/resources/security/RealmResource.java
@@ -14,7 +14,6 @@
  * limitations under the License.
  */
 
-
 package org.slc.sli.api.resources.security;
 
 import java.util.ArrayList;
@@ -63,10 +62,10 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Component;
 
 /**
- * Realm role mapping API. Allows full CRUD on realm objects. Primarily intended to allow
- * mappings between SLI roles and client roles as realms should not be created or deleted
- * frequently.
- *
+ * Realm role mapping API. Allows full CRUD on realm objects. Primarily intended
+ * to allow mappings between SLI roles and client roles as realms should not be
+ * created or deleted frequently.
+ * 
  * @author jnanney
  */
 @Component
@@ -98,17 +97,10 @@ public class RealmResource {
     @Autowired
     private RoleInitializer roleInitializer;
 
-
-
     @PostConstruct
     public void init() {
         EntityDefinition def = store.lookupByResourceName(REALM);
         setService(def.getService());
-    }
-
-    // Injector
-    public void setStore(EntityDefinitionStore store) {
-        this.store = store;
     }
 
     // Injector
@@ -119,7 +111,7 @@ public class RealmResource {
     @PUT
     @Path("{realmId}")
     @Consumes("application/json")
-    @RightsAllowed({Right.CRUD_REALM })
+    @RightsAllowed({ Right.CRUD_REALM })
     public Response updateRealm(@PathParam("realmId") String realmId, EntityBody updatedRealm,
             @Context final UriInfo uriInfo) {
 
@@ -135,7 +127,8 @@ public class RealmResource {
             body.put(RESPONSE, "You are not authorized to update this realm.");
             return Response.status(Status.FORBIDDEN).entity(body).build();
         }
-        Response validateUniqueness = validateUniqueId(realmId, (String) updatedRealm.get(UNIQUE_IDENTIFIER), (String) updatedRealm.get(NAME));
+        Response validateUniqueness = validateUniqueId(realmId, (String) updatedRealm.get(UNIQUE_IDENTIFIER),
+                (String) updatedRealm.get(NAME));
         if (validateUniqueness != null) {
             return validateUniqueness;
         }
@@ -145,8 +138,8 @@ public class RealmResource {
         updatedRealm.put(ED_ORG, SecurityUtil.getEdOrg());
 
         if (service.update(realmId, updatedRealm)) {
-            audit(securityEventBuilder.createSecurityEvent(RealmResource.class.getName(), uriInfo.getRequestUri(), "Realm ["
-                    + updatedRealm.get(NAME) + "] updated!"));
+            audit(securityEventBuilder.createSecurityEvent(RealmResource.class.getName(), uriInfo.getRequestUri(),
+                    "Realm [" + updatedRealm.get(NAME) + "] updated!"));
             return Response.status(Status.NO_CONTENT).build();
         }
         return Response.status(Status.BAD_REQUEST).build();
@@ -154,18 +147,23 @@ public class RealmResource {
 
     @DELETE
     @Path("{realmId}")
-    @RightsAllowed({Right.CRUD_REALM })
+    @RightsAllowed({ Right.CRUD_REALM })
     public Response deleteRealm(@PathParam("realmId") String realmId, @Context final UriInfo uriInfo) {
         EntityBody deletedRealm = service.get(realmId);
+        if (!canEditCurrentRealm(deletedRealm)) {
+            EntityBody body = new EntityBody();
+            body.put(RESPONSE, "You are not authorized to delete a realm for another ed org");
+            return Response.status(Status.FORBIDDEN).entity(body).build();
+        }
         service.delete(realmId);
         roleInitializer.dropRoles(realmId);
-        audit(securityEventBuilder.createSecurityEvent(RealmResource.class.getName(), uriInfo.getRequestUri(), "Realm ["
-                + deletedRealm.get(NAME) + "] deleted!"));
+        audit(securityEventBuilder.createSecurityEvent(RealmResource.class.getName(), uriInfo.getRequestUri(),
+                "Realm [" + deletedRealm.get(NAME) + "] deleted!"));
         return Response.status(Status.NO_CONTENT).build();
     }
 
     @POST
-    @RightsAllowed({Right.CRUD_REALM })
+    @RightsAllowed({ Right.CRUD_REALM })
     public Response createRealm(EntityBody newRealm, @Context final UriInfo uriInfo) {
 
         if (!canEditCurrentRealm(newRealm)) {
@@ -173,7 +171,8 @@ public class RealmResource {
             body.put(RESPONSE, "You are not authorized to create a realm for another ed org");
             return Response.status(Status.FORBIDDEN).entity(body).build();
         }
-        Response validateUniqueness = validateUniqueId(null, (String) newRealm.get(UNIQUE_IDENTIFIER), (String) newRealm.get(NAME));
+        Response validateUniqueness = validateUniqueId(null, (String) newRealm.get(UNIQUE_IDENTIFIER),
+                (String) newRealm.get(NAME));
         if (validateUniqueness != null) {
             debug("On realm create, uniqueId is not unique");
             return validateUniqueness;
@@ -185,11 +184,11 @@ public class RealmResource {
 
         String id = service.create(newRealm);
 
-        //Also create custom roles
+        // Also create custom roles
         roleInitializer.dropAndBuildRoles(id);
 
-        audit(securityEventBuilder.createSecurityEvent(RealmResource.class.getName(), uriInfo.getRequestUri(), "Realm ["
-                + newRealm.get(NAME) + "] created!"));
+        audit(securityEventBuilder.createSecurityEvent(RealmResource.class.getName(), uriInfo.getRequestUri(),
+                "Realm [" + newRealm.get(NAME) + "] created!"));
         String uri = uriToString(uriInfo) + "/" + id;
 
         return Response.status(Status.CREATED).header("Location", uri).build();
@@ -197,19 +196,20 @@ public class RealmResource {
 
     @GET
     @Path("{realmId}")
-    @RightsAllowed({Right.ADMIN_ACCESS })
+    @RightsAllowed({ Right.ADMIN_ACCESS })
     public Response readRealm(@PathParam("realmId") String realmId) {
         EntityBody result = service.get(realmId);
         return Response.ok(result).build();
     }
 
     @GET
-    @RightsAllowed({Right.ADMIN_ACCESS })
+    @RightsAllowed({ Right.ADMIN_ACCESS })
     public Response getRealms(@QueryParam(REALM) @DefaultValue("") String realm, @Context UriInfo info) {
         SLIPrincipal principal = (SLIPrincipal) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
 
         NeutralQuery neutralQuery = new NeutralQuery();
-        neutralQuery.addCriteria(new NeutralCriteria("tenantId", NeutralCriteria.OPERATOR_EQUAL, principal.getTenantId()));
+        neutralQuery.addCriteria(new NeutralCriteria("tenantId", NeutralCriteria.OPERATOR_EQUAL, principal
+                .getTenantId()));
         neutralQuery.addCriteria(new NeutralCriteria(ED_ORG, NeutralCriteria.OPERATOR_EQUAL, principal.getEdOrg()));
         neutralQuery.setOffset(0);
         neutralQuery.setLimit(100);
@@ -235,7 +235,6 @@ public class RealmResource {
         return Response.ok(result).build();
     }
 
-
     private Response validateUniqueId(String realmId, String uniqueId, String displayName) {
         if (uniqueId == null || uniqueId.length() == 0) {
             return null;
@@ -246,13 +245,13 @@ public class RealmResource {
         if (realmId != null) {
             query.addCriteria(new NeutralCriteria("_id", "!=", idConverter.toDatabaseId(realmId)));
         }
-        Entity body =
-                SecurityUtil.runWithAllTenants(new SecurityTask<Entity>() {
+        Entity body = SecurityUtil.runWithAllTenants(new SecurityTask<Entity>() {
 
-                    @Override
-                    public Entity execute() {
-                        return repo.findOne(REALM, query);
-                    } });
+            @Override
+            public Entity execute() {
+                return repo.findOne(REALM, query);
+            }
+        });
 
         if (body != null) {
             debug("uniqueId: {}", body.getBody().get(UNIQUE_IDENTIFIER));
